little improvements
most of them are renaming
