+++ AntHill/ClassLibraryAntHill/AntHill.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Drawing;
+
+namespace ClassLibraryAntHill
+{
+    public class AntHill
+    {
+        PointF center;
+    }
+}
