+++ AntHill/ClassLibraryAntHill/Field.cs
@@ -7,133 +7,7 @@ namespace ClassLibraryAntHill
 {
 
     public class Field
-    {
-        public class AntBuilder : Ant
-        {
-            private const int speed = 3;
-            private List<Food> OpenFoods;
-            public bool Bring { get; private set; }
-            public AntBuilder(double x, double y, string name) : base(x, y, name)
-            {
-                OpenFoods = new List<Food>();
-                Bring = false;
-            }
-            public override void BeAttak()
-            {
-            }
-            public void GiveOpenFoods(List<Food> food)
-            {
-                OpenFoods = food;
-            }
-            public override int Thinking()
-            {
-                int index = MinDictance();
-                if (commands[0].action == Action.findfood)
-                {
-                    if (index != -1)
-                    {
-                        commands.RemoveAll(x => true);
-                        commands.Add(new Command(Action.gotoFood, OpenFoods[index].X, OpenFoods[index].Y));
-                    }
-                    else
-                    {
-                        if (X > 10 && X < 650 && Y > 10 && Y < 400)
-                        {
-                            double a =Math.Atan((LastY - Y)/(LastX - X));
-                            Move(Math.Cos(a)*speed, Math.Sin(a)*speed);
-                            return -2;
-                        }
-                        else
-                        {
-                            commands.Clear();
-                            commands.Add(new Command(Action.gotoAntHill, 100, 100));
-                            double a = Math.Atan((100 - Y) / (100 - X));
-                            Move(Math.Cos(a), Math.Sin(a));
-                        }
-                        
-                    }
-                }
-
-                if (commands[0].action == Action.gotoFood)
-                {
-                    if (index == -1)
-                    {
-                        commands.Clear();
-                        commands.Add(new Command(Action.findfood));
-                    }
-                    else
-                    {
-                        double d = Math.Sqrt((X - OpenFoods[index].X) * (X - OpenFoods[index].X) + (Y - OpenFoods[index].Y) * (Y - OpenFoods[index].Y));
-                        if (d < 10)
-                        {
-                            commands.Clear();
-                            commands.Add(new Command(Action.gotoAntHill, 100, 100));
-                            double a = Math.Atan((OpenFoods[index].Y - Y) / (OpenFoods[index].X - X));
-                            Move(Math.Cos(a), Math.Sin(a));
-                            Bring = true;
-                            OpenFoods[index].ChangeFood();
-                        }
-                        else
-                        {
-                            double a = Math.Atan((OpenFoods[index].Y - Y) / (OpenFoods[index].X - X));
-                            double al = AntMath.GetAnkleBetwentwoVector(-OpenFoods[index].X + X, -OpenFoods[index].Y + Y, X - LastX, Y - LastY);
-                            if (Math.Abs(al) < Math.PI / 2)
-                            {
-                                Move(Math.Cos(a), Math.Sin(a));
-                                return -1;
-                            }
-
-                                Move( Math.Cos(a) * speed, Math.Sin(a) * speed);
-    
-                        }
-                    }
-                }
-                if (commands[0].action == Action.gotoAntHill)
-                {
-                    double d = Math.Sqrt((X - 100) * (X - 100) + (Y - 100) * (Y - 100));
-                    if (d < 5)
-                    {
-                        Bring = false;
-                        commands.RemoveAll(x => true);
-                        commands.Add(new Command(Action.findfood));
-                    }
-                    else
-                    {
-                        double a = Math.Atan((100 - Y)/(100 - X));
-                        double al = AntMath.GetAnkleBetwentwoVector( X - 100 ,  Y - 100 , X - LastX, Y - LastY);
-                        if (Math.Abs(al) < Math.PI / 2)
-                        {
-                            Move(-Math.Cos(a), -Math.Sin(a));
-                            return -2;
-                        }
- 
-                            Move(- Math.Cos(a) * speed,-Math.Sin(a) * speed);
-
-                        return -2;
-                    }
-                }
-                if (!(X > 10 && X < 650 && Y > 10 && Y < 400))
-                {
-                    bool ok = true;
-                }
-                    return -1;
-            }
-            private int MinDictance()
-            {
-                double dist = 400;
-                int index = -1;
-                for(int i=0;i<OpenFoods.Count;i++)
-                {
-                    double d = Math.Sqrt((X - OpenFoods[i].X) * (X - OpenFoods[i].X) + (Y - OpenFoods[i].Y) * (Y - OpenFoods[i].Y));
-                    if(dist>d)
-                    {
-                        dist = d;
-                        index = i;
-                    }
-                }
-                return index;
-            }
-        }
+    {        
         public List<Ant> Ants { get; private set; }
         public List<Ant> Pests { get; private set; }
         public List<Food> Foods { get; private set; }
@@ -147,13 +21,13 @@ namespace ClassLibraryAntHill
         }
         public void BornAnt(double x,double y,string name)
         {
-            Ants.Add(new AntBuilder(x, y,name));
+            Ants.Add(new WorkerAnt(x, y,name,null));
         }
         public void BornFood(double x, double y)
         {
             Foods.Add(new Food(x, y));
         }
-        private void Alive()
+        private void RemoveDead()
         {
             Ants.RemoveAll(x => x.Hp <= 0);
             Pests.RemoveAll(x => x.Hp <= 0);
@@ -169,7 +43,7 @@ namespace ClassLibraryAntHill
                     Ants[i].commands.Add(new Command(Action.findfood));
                 }
 
-                ((AntBuilder)Ants[i]).GiveOpenFoods(OpenFoods);
+                ((WorkerAnt)Ants[i]).GiveOpenFoods(OpenFoods);
                 int h = Ants[i].Thinking();
                 if (h == -2)
                 {
@@ -187,7 +61,7 @@ namespace ClassLibraryAntHill
                 }
 
             }
-            Alive();
+            RemoveDead();
         }
     }
 }
